final ok

+++ Project_All/profile.php
@@ -3,7 +3,7 @@
 <?php include('labor/dbConfig.php');
 	session_start();
 	if(isset($_SESSION['emailname'])){
-		echo "welcome".$_SESSION['emailname'];
+		// echo "welcome".$_SESSION['emailname'];
 }
 else
 {
@@ -131,6 +131,7 @@ if(isset($_POST['logout']))
 
 <!-- customer update profile start -->
 	<?php
+		    $flag='';
 	    if(isset($_POST['updateprofile']))
 		{
 
@@ -146,7 +147,7 @@ if(isset($_POST['logout']))
 		      $pincode=$_POST['pincode'];
 		      $password=$_POST['password'];
 		      $about=$_POST['about'];
-		      $flag='';
+		     
 		      
 		      	if($_FILES["fimage"]["name"]!=null)
 		      	{
@@ -157,10 +158,14 @@ if(isset($_POST['logout']))
 	     		  while($row=mysqli_fetch_row($res))
 				  {
 				  	$fimage=$row[13];
-				  }
+				  	if(!empty($row[13]))
+				  	{
 
-			      $oldimage=$path.$fimage;
-			      unlink($oldimage);
+			      		$oldimage=$path.$fimage;
+			      		unlink($oldimage);
+				  	}
+
+				  }
 
 			      $target_file = $path.basename($_FILES["fimage"]["name"]);
 			      // Select file type
@@ -216,7 +221,8 @@ if(isset($_POST['logout']))
 		        if($res>0)
 		        {
 		          // echo "update record into customer table";
-
+					echo "<script> window.location.href='profile.php';</script>";
+				   	exit;
 		         // header("location:index.php");
 		        }   
 		        else
@@ -570,7 +576,7 @@ if(isset($_POST['logout']))
 									{
 									$stateid=$row1[0];
 										?>
-	 					            <li><a class="justify-content-between d-flex" href="category.php">
+	 					            <li><a class="justify-content-between d-flex" href="search.php?stateid=<?php echo $row1[0]?>">
 	 								<p><?php echo $row1[1];?></p>
 									<?php
 									$qry2="select count(*) as sta from labor where l_dflag<>'1' and l_state='$stateid' group by l_state";
@@ -594,7 +600,7 @@ if(isset($_POST['logout']))
 								</ul>
 							</div>
 
-							<div class="single-slidebar">
+							<!-- <div class="single-slidebar">
 								<h4>Top rated labor</h4>
 								<div class="active-relatedjob-carusel">
 								
@@ -638,7 +644,7 @@ if(isset($_POST['logout']))
 									</div>
 									 <?php
 										}
-							        ?>
+							        ?> -->
 
 
 								<!-- 	<div class="single-rated">
@@ -677,8 +683,8 @@ if(isset($_POST['logout']))
 										<p class="address"><span class="lnr lnr-database"></span> 15k - 25k</p>
 										<a href="#" class="btns text-uppercase">Apply job</a>
 									</div> -->																		
-								</div>
-							</div>							
+								<!-- </div>
+							</div>	 -->						
 
 							<div class="single-slidebar">
 								<h4>Labor by Category</h4>
@@ -691,7 +697,7 @@ if(isset($_POST['logout']))
 									{
 									$catid=$row1[0];
 										?>
-	 					            <li><a class="justify-content-between d-flex" href="category.php">
+	 					            <li><a class="justify-content-between d-flex"  href="category.php?caid=<?php echo $row1[0]; ?>">
 	 								<p><?php echo $row1[1];?></p>
 									<?php
 									$qry2="select count(*) as cat from labor where l_dflag<>'1' and l_categoryid='$catid' group by l_categoryid";
@@ -713,6 +719,42 @@ if(isset($_POST['logout']))
 									<li><a class="justify-content-between d-flex" href="category.html"><p>Accounting</p><span>17</span></a></li> -->
 								</ul>
 							</div>
+
+							<div class="single-slidebar">
+								<h4>Total labor under leader</h4>
+								<ul class="cat-list">
+
+									<?php
+									$qry6="select * from labor where l_leaderid<>'' and l_dflag<>'1' group by l_leaderid";
+									$res6=mysqli_query($con,$qry6);
+									while($row6=mysqli_fetch_row($res6))
+									{
+										$ladid=$row6[21];
+
+										$qry8="select * from labor where l_id='$row6[21]'";
+										$res8=mysqli_query($con,$qry8);
+										while($row8=mysqli_fetch_row($res8))
+										{
+											$leadername=$row8[1].' '.$row8[2];
+										}
+									
+										?>
+	 					      		<li><a class="justify-content-between d-flex" href="single.php?lid=<?php echo $row6[21]; ?>">
+	 								<?php echo $leadername;?>
+									<?php
+									$qry7="select count(*) as led from labor where l_dflag<>'1' and l_leaderid='$ladid' group by l_leaderid";
+									// echo $qry
+									$res7=mysqli_query($con,$qry7);
+									while($row7=mysqli_fetch_array($res7))
+									{
+									?>
+            						<span><?php echo $row7['led']; ?></span></a></li>
+						            <?php
+	          							}
+	          						}?>
+	          					</ul>
+							</div>
+
 
 <!-- 							<div class="single-slidebar">
 								<h4>Carrer Advice Blog</h4>
